Move ConfigBase to plugin module
