+++ even-more-fish-api/src/main/java/com/oheers/fish/api/config/ConfigUtils.java
@@ -1,6 +1,6 @@
 package com.oheers.fish.api.config;
 
-import dev.dejvokep.boostedyaml.block.implementation.Section;
+import org.bukkit.configuration.ConfigurationSection;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
 
@@ -9,8 +9,8 @@ import org.jspecify.annotations.Nullable;
  */
 public class ConfigUtils {
 
-    public static @NonNull Section getOrCreateSection(@NonNull Section section, @NonNull String path) {
-        Section subSection = section.getSection(path);
+    public static @NonNull ConfigurationSection getOrCreateSection(@NonNull ConfigurationSection section, @NonNull String path) {
+        ConfigurationSection subSection = section.getConfigurationSection(path);
         if (subSection == null) {
             subSection = section.createSection(path);
         }
@@ -21,9 +21,9 @@ public class ConfigUtils {
      * Gets the first section of many paths.
      * Useful for typos... Oops
      */
-    public static @Nullable Section getSectionOfMany(@NonNull Section section, @NonNull String... paths) {
+    public static @Nullable ConfigurationSection getSectionOfMany(@NonNull ConfigurationSection section, @NonNull String... paths) {
         for (String path : paths) {
-            Section sub = section.getSection(path);
+            ConfigurationSection sub = section.getConfigurationSection(path);
             if (sub != null) {
                 return sub;
             }
