+++ even-more-fish-plugin/src/main/java/com/oheers/fish/baits/BaitHandler.java
@@ -4,7 +4,7 @@ import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
 import com.oheers.fish.api.Logging;
 import com.oheers.fish.api.baits.IBait;
-import com.oheers.fish.api.config.ConfigBase;
+import com.oheers.fish.config.ConfigBase;
 import com.oheers.fish.api.economy.Economy;
 import com.oheers.fish.api.economy.EconomyType;
 import com.oheers.fish.api.fishing.FishingType;
