+++ even-more-fish-plugin/src/main/java/com/oheers/fish/baits/configs/BaitConversions.java
@@ -2,7 +2,7 @@ package com.oheers.fish.baits.configs;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.api.Logging;
-import com.oheers.fish.api.config.ConfigBase;
+import com.oheers.fish.config.ConfigBase;
 import com.oheers.fish.config.MainConfig;
 import com.oheers.fish.config.MessageConfig;
 import dev.dejvokep.boostedyaml.YamlDocument;
