+++ even-more-fish-plugin/src/main/java/com/oheers/fish/fishing/items/config/FishConversions.java
@@ -2,7 +2,7 @@ package com.oheers.fish.fishing.items.config;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.api.Logging;
-import com.oheers.fish.api.config.ConfigBase;
+import com.oheers.fish.config.ConfigBase;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
 import org.jspecify.annotations.NonNull;
 import uk.firedev.daisylib.messages.message.ComponentMessage;
